+++ src/com/wywy/log4j/appender/FluencyAppender.java
@@ -22,19 +22,22 @@ import java.util.Map;
 @Plugin(name="Fluency", category="Core", elementType="appender", printObject=true)
 public final class FluencyAppender extends AbstractAppender {
 
-    private final org.slf4j.Logger LOG = LoggerFactory.getLogger(getClass());
+    private static final org.slf4j.Logger LOG = LoggerFactory.getLogger(FluencyAppender.class);
     private final SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
     private final NameAbbreviator abbr = NameAbbreviator.getAbbreviator("1.");
 
     private Fluency fluency;
     private Map<String, Object> parameters;
+    private Map<String, String> staticFields;
 
-    private FluencyAppender(final String name, final Map<String, Object> parameters, final Filter filter, final Layout<? extends Serializable> layout,
+    private FluencyAppender(final String name, final Map<String, Object> parameters, final Map<String, String> staticFields,
+                            final Filter filter, final Layout<? extends Serializable> layout,
                             final boolean ignoreExceptions) {
 
         super(name, filter, layout, ignoreExceptions);
 
         this.parameters = parameters;
+        this.staticFields = staticFields;
 
         // Single Fluentd(localhost:24224 by default)
         //   - TCP heartbeat (by default)
@@ -49,6 +52,7 @@ public final class FluencyAppender extends AbstractAppender {
         //   - Max wait until the flusher is terminated is 10 seconds (by default)
         try {
             fluency = Fluency.defaultFluency();
+            LOG.info("FluencyAppender initialized");
         } catch (IOException e) {
             LOG.error(e.getMessage());
         }
@@ -59,11 +63,13 @@ public final class FluencyAppender extends AbstractAppender {
                                                  @PluginAttribute("tag") final String tag,
                                                  @PluginAttribute("application") final String application,
                                                  @PluginAttribute("ignoreExceptions") final String ignore,
+                                                 @PluginElement("StaticField") final StaticField[] staticFields,
                                                  @PluginElement("Layout") Layout<? extends Serializable> layout,
                                                  @PluginElement("Filter") final Filter filter) {
         final boolean ignoreExceptions = Booleans.parseBoolean(ignore, true);
 
         Map<String, Object> parameters = new HashMap<>();
+        Map<String, String> fields = new HashMap<>();
 
         // the @Required annotation for Attributes just returns a non helpful exception, so i'll check it manually
         // we need the tag for fluency itself (it's actually a part of the fluentd protocol
@@ -73,11 +79,24 @@ public final class FluencyAppender extends AbstractAppender {
             throw new IllegalArgumentException("tag is required");
         }
 
+        // Deprecated
         if (application != null) {
-            parameters.put("application", application);
+            fields.put("application", application);
         }
 
-        return new FluencyAppender(name, parameters, filter, layout, ignoreExceptions);
+        for (StaticField field: staticFields) {
+            if (field.getName().trim().equals("")) {
+                LOG.warn("Skipping empty field");
+                continue;
+            }
+            if (field.getValue().trim().equals("")) {
+                LOG.warn("Skipping field {} due to empty value", field.getName());
+                continue;
+            }
+            fields.put(field.getName(), field.getValue());
+        }
+
+        return new FluencyAppender(name, parameters, fields, filter, layout, ignoreExceptions);
     }
 
     @Override
@@ -116,13 +135,11 @@ public final class FluencyAppender extends AbstractAppender {
             m.put("sourceLine", logEvent.getSource().getLineNumber());
         }
 
-        if (parameters.containsKey("application")) {
-            m.put("application", parameters.get("application"));
-        }
         m.put("logger", abbr.abbreviate(loggerName));
         m.put("loggerFull", loggerName);
         m.put("message", message);
         m.put("thread", logEvent.getThreadName());
+        m.putAll(this.staticFields);
 
         // TODO: get rid of that, once the whole stack supports subsecond timestamps
         // this is just a workaround due to the lack of support