+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.1.0
+* add generic `<StaticField>` configuration
+    * deprecates `application=""` attribute (it still works, but please migrate)
 ## 1.0.0
 * switch groupId for maven central registration